+++ CHANGELOG.rst
@@ -17,6 +17,7 @@ Added
 - ``Sequence.seq_encoded`` to allow turning a sequence into an encoded list
 - ``Sequence.encoded_matrix`` to give the entire alignment as encoded matrix
 - ``SequenceFile.filter_gapped`` to filter sequences with a certain threshold of gaps
+- ``SequenceFile.to_string`` and ``ContactMap.to_string`` methods
 Changed
 ~~~~~~~
 - Changed API interface for ``conkit.plot`` in accordance to necessary changes for above